Enable -Wextra and fix issues (#12591)
This enables `-Wextra`, removes warning flags already covered by it in
GCC and Clang, and fixes a comparison between integers of different
signs as well as redundant moves in initializations.

+++ src/lib/ffs.c
@@ -21,7 +21,7 @@ extern "C" {
 
 int ffs(int i) {
   long mask = 0x1;
-  int pos = 1;
+  size_t pos = 1;
   while(pos <= sizeof(int) * 8) {
     if((mask & i) != 0) {
       return pos;
